Changed file extension + vim modeline
In order to allow standard tools to understand that the source code is
in assembly (for example, in the case of Github's language detection), I
decided to rename the file extension from .agb to .asm. In any case, vim
wouldn't have gotten it right anyway (since the filetype needs to
specifically be rgbasm), so I also added a vim modeline to the bottom of
the file.