Fixed a major bug (one character!)
When following a tutorial, I made a modification to the code, but got it
wrong accidentally. mednafen was running the program just fine, but
VisualBoy Advance had extremely messed up graphics. I assumed it was
some inaccuracy in VBA, but found that gambatte, an accurate emulator,
had the same exact problem, so I finally fixed it.

I added some code to clear out the OAM completely, and updated some
documentation to make sure VisualBoy Advance was mentioned as working,
and mentioned gambatte.