Added walking animation
When the player moves, the sprite animates through four frames, but not
at one frame per V-Blank. Instead, the rate is slower and more natural.
Additionally, if the player stops moving but the sprite is in the middle
of the walk cycle animation, the animation continues until the end of
the cycle.