Added stat-interrupt subproject
Shows the LCDC Status interrupt, namely the H-Blank interrupt, which is
used to perform a fun little video effect. Additionally, the LYC
register and corresponding flag in the LCDC Status register are used to
show the window at only the top and bottom of the screen.

There is some flickering that needs to be taken care of.