Store player state within struct in memory
This allows us to manipulate these values during each frame. Actually,
the X- and Y-coordinates are easy enough to read from the OAM buffer,
but, for example, the player consists of two sprites with two different
X-coordinates. It's easier to have a single, canonical location for the
"overall" player X position.