Start of checking for ground collision
If the player is exclusively over a single column, then the logic is
implemented to correctly find the height the player should be placed.
For debugging purposes, when the player is partially over two columns,
the player falls to the bottom of the screen.

The ground-checking logic needs to be refactored so there's not so much
duplication. This will be especially important when checking multiple
columns, as the logic shouldn't be duplicated once for each column
checked!