Add "buttons" subproject
A simple test program that shows which button (or multiple buttons) is
currently being pressed. This functionality allows testing controllers
on emulators.

Includes a sample screenshot showing what the screen would look like
with all the buttons pressed. This is mainly to show all available
images that may at one point be visible. Serves as a good reference
point in case I want to tweak the assets in the future.